0-print_dlistint.c

+++ doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t cpt = 0;
+	unsigned int cpt = 0;
 	const dlistint_t *current = h;
 
 	while (current != NULL)
